First scanner draft. Still have unsolved publishing issue
Tries to keep an open model (does not know the number of fields in the
csv file, nor their type).
But for now, fails at the publishing to DB stage : even when getting the
DB fields names, it doesn't seem to suffice : you can't just send
strings and hope they will be properly cast.