Rewritten the INSERT (-only) statement
It is now possible to choose in properties if the records are updatable
or not (general setup). It can also be done in the files naming : if no
key are listed, the data in the table is assumed to be non-updatable
(per-file choice).
The INSERT procedure is now similar to the UPSERT one (same parsing
functions, more resilient than using PreparedStatement)